Revert "Move body buffer copy from before phyics to before render call. This allows it to be skipped when drawing is disabled while still providing a double-buffer benefit."
This reverts commit 14743a8f08ec127f68a971cbdf07e4230c59d211.

+++ NBodies/UI/Display.cs
@@ -729,7 +729,7 @@ namespace NBodies.UI
 
         private void PressureScaleUpDown_ValueChanged(object sender, EventArgs e)
         {
-            RenderBase.StyleScaleMax = (int)PressureScaleUpDown.Value;
+            RenderBase.StyleScaleMax = (float)PressureScaleUpDown.Value;
         }
 
         private void CenterOnMassButton_Click(object sender, EventArgs e)
